refactor things a bit to make boundaries a bit more clear
