+++ playground/react-server/package.json
@@ -22,7 +22,10 @@
 		"@types/react-dom": "^19.0.0",
 		"@vite-plugin-cloudflare/typescript-config": "workspace:*",
 		"@vitejs/plugin-react": "^4.3.4",
+		"autoprefixer": "^10.4.20",
+		"postcss": "^8.4.49",
 		"rsc-html-stream": "0.0.4",
+		"tailwindcss": "^3.4.17",
 		"typescript": "catalog:default",
 		"unenv": "catalog:default",
 		"unplugin-rsc": "0.0.11",
