+++ playground/react-server/src/app/app.tsx
@@ -1,4 +1,4 @@
-import { getEnv, getURL } from './context';
+import { getEnv, getURL } from '../framework/server';
 import { Counter } from './counter';
 
 export async function App() {
